Move config files to config folder
If we have config.json and index.js at the root, it seems to cause
problems with nuxt
